Moved fetch_pdb() to PDBParser
