+++ package/MDAnalysis/__init__.py
@@ -221,6 +221,9 @@ from . import converters
 
 from .due import due, Doi, BibTeX
 
+## Temp fetch_pdb import to see if things works
+from .topology.PDBParser import fetch_pdb
+
 due.cite(
     Doi("10.25080/majora-629e541a-00e"),
     description="Molecular simulation analysis library",
